ai/models/threat_detector.py
import torch
import torch.nn as nn
import torch.nn.functional as F
from torchvision.models import resnet50
from transformers import AutoModel, AutoTokenizer
import numpy as np
from typing import Dict, List, Tuple, Optional
import logging

logger = logging.getLogger(__name__)

class MultiModalAttention(nn.Module):
    """Advanced attention mechanism for multi-modal fusion"""
    def __init__(self, dim: int, num_heads: int = 8):
        super().__init__()
        self.num_heads = num_heads
        self.scale = (dim // num_heads) ** -0.5
        self.qkv = nn.Linear(dim, dim * 3)
        self.proj = nn.Linear(dim, dim)
        
    def forward(self, x: torch.Tensor) -> torch.Tensor:
        B, N, C = x.shape
        qkv = self.qkv(x).reshape(B, N, 3, self.num_heads, C // self.num_heads)
        qkv = qkv.permute(2, 0, 3, 1, 4)
        q, k, v = qkv[0], qkv[1], qkv[2]
        
        attn = (q @ k.transpose(-2, -1)) * self.scale
        attn = attn.softmax(dim=-1)
        
        x = (attn @ v).transpose(1, 2).reshape(B, N, C)
        x = self.proj(x)
        return x

class ThreatDetector(nn.Module):
    """Advanced threat detection model with multi-modal fusion"""
    def __init__(
        self,
        video_backbone: str = "resnet50",
        audio_backbone: str = "wav2vec2-base",
        num_threat_classes: int = 10,
        fusion_dim: int = 512
    ):
        super().__init__()
        
        # Video processing
        self.video_encoder = resnet50(pretrained=True)
        self.video_encoder.fc = nn.Linear(2048, fusion_dim)
        
        # Audio processing
        self.audio_encoder = AutoModel.from_pretrained(f"facebook/{audio_backbone}")
        self.audio_proj = nn.Linear(768, fusion_dim)
        
        # Multi-modal fusion
        self.fusion_attention = MultiModalAttention(fusion_dim)
        self.temporal_conv = nn.Conv1d(fusion_dim, fusion_dim, kernel_size=3, padding=1)
        
        # Threat classification
        self.classifier = nn.Sequential(
            nn.Linear(fusion_dim, 256),
            nn.ReLU(),
            nn.Dropout(0.2),
            nn.Linear(256, num_threat_classes)
        )
        
        # Threat severity estimation
        self.severity_estimator = nn.Sequential(
            nn.Linear(fusion_dim, 128),
            nn.ReLU(),
            nn.Linear(128, 1),
            nn.Sigmoid()
        )
        
        # Initialize weights
        self._init_weights()
        
    def _init_weights(self):
        """Initialize model weights using advanced techniques"""
        for m in self.modules():
            if isinstance(m, nn.Linear):
                nn.init.kaiming_normal_(m.weight)
                if m.bias is not None:
                    nn.init.zeros_(m.bias)
            elif isinstance(m, nn.Conv1d):
                nn.init.kaiming_normal_(m.weight)
                if m.bias is not None:
                    nn.init.zeros_(m.bias)
    
    def forward(
        self,
        video: torch.Tensor,
        audio: torch.Tensor,
        return_attention: bool = False
    ) -> Dict[str, torch.Tensor]:
        """
        Forward pass with multi-modal fusion
        
        Args:
            video: Video tensor of shape (B, C, H, W)
            audio: Audio tensor of shape (B, T)
            return_attention: Whether to return attention maps
            
        Returns:
            Dictionary containing threat predictions and attention maps
        """
        # Process video
        video_features = self.video_encoder(video)
        
        # Process audio
        audio_outputs = self.audio_encoder(audio)
        audio_features = self.audio_proj(audio_outputs.last_hidden_state)
        
        # Fuse modalities
        combined_features = torch.stack([video_features, audio_features.mean(dim=1)], dim=1)
        fused_features = self.fusion_attention(combined_features)
        
        # Temporal processing
        temporal_features = self.temporal_conv(fused_features.transpose(1, 2)).transpose(1, 2)
        
        # Threat classification
        threat_logits = self.classifier(temporal_features.mean(dim=1))
        threat_probs = F.softmax(threat_logits, dim=-1)
        
        # Severity estimation
        severity = self.severity_estimator(temporal_features.mean(dim=1))
        
        outputs = {
            'threat_probs': threat_probs,
            'severity': severity,
            'features': temporal_features
        }
        
        if return_attention:
            outputs['attention_maps'] = self.fusion_attention.attn
            
        return outputs
    
    def predict_threat(
        self,
        video: torch.Tensor,
        audio: torch.Tensor,
        threshold: float = 0.8
    ) -> Dict[str, np.ndarray]:
        """
        Predict threats with confidence scores
        
        Args:
            video: Video tensor
            audio: Audio tensor
            threshold: Confidence threshold for threat detection
            
        Returns:
            Dictionary containing threat predictions and confidence scores
        """
        self.eval()
        with torch.no_grad():
            outputs = self.forward(video, audio)
            
            # Get predictions
            threat_probs = outputs['threat_probs'].cpu().numpy()
            severity = outputs['severity'].cpu().numpy()
            
            # Apply threshold
            threat_detected = threat_probs.max(axis=1) > threshold
            
            return {
                'threat_detected': threat_detected,
                'threat_class': threat_probs.argmax(axis=1),
                'confidence': threat_probs.max(axis=1),
                'severity': severity.squeeze()
            }
    
    def train_step(
        self,
        video: torch.Tensor,
        audio: torch.Tensor,
        labels: torch.Tensor,
        severity_labels: torch.Tensor
    ) -> Dict[str, float]:
        """
        Training step with advanced loss functions
        
        Args:
            video: Video tensor
            audio: Audio tensor
            labels: Threat class labels
            severity_labels: Threat severity labels
            
        Returns:
            Dictionary containing loss values
        """
        self.train()
        
        # Forward pass
        outputs = self.forward(video, audio)
        
        # Calculate losses
        classification_loss = F.cross_entropy(outputs['threat_probs'], labels)
        severity_loss = F.mse_loss(outputs['severity'], severity_labels)
        
        # Feature consistency loss
        feature_consistency_loss = self._compute_feature_consistency(outputs['features'])
        
        # Total loss
        total_loss = (
            classification_loss +
            0.5 * severity_loss +
            0.1 * feature_consistency_loss
        )
        
        return {
            'total_loss': total_loss.item(),
            'classification_loss': classification_loss.item(),
            'severity_loss': severity_loss.item(),
            'consistency_loss': feature_consistency_loss.item()
        }
    
    def _compute_feature_consistency(self, features: torch.Tensor) -> torch.Tensor:
        """Compute feature consistency loss for temporal smoothness"""
        return F.mse_loss(features[:, 1:], features[:, :-1])

class ThreatAnalyzer:
    """High-level threat analysis interface"""
    def __init__(
        self,
        model_path: Optional[str] = None,
        device: str = "cuda" if torch.cuda.is_available() else "cpu"
    ):
        self.device = device
        self.model = ThreatDetector().to(device)
        
        if model_path:
            self.load_model(model_path)
    
    def load_model(self, path: str):
        """Load trained model weights"""
        try:
            state_dict = torch.load(path, map_location=self.device)
            self.model.load_state_dict(state_dict)
            logger.info(f"Loaded model from {path}")
        except Exception as e:
            logger.error(f"Failed to load model: {e}")
            raise
    
    def analyze_threat(
        self,
        video: np.ndarray,
        audio: np.ndarray
    ) -> Dict[str, np.ndarray]:
        """
        Analyze video and audio for threats
        
        Args:
            video: Video array of shape (H, W, C)
            audio: Audio array of shape (T,)
            
        Returns:
            Dictionary containing threat analysis results
        """
        # Preprocess inputs
        video_tensor = self._preprocess_video(video)
        audio_tensor = self._preprocess_audio(audio)
        
        # Get predictions
        predictions = self.model.predict_threat(video_tensor, audio_tensor)
        
        # Post-process results
        return self._postprocess_predictions(predictions)
    
    def _preprocess_video(self, video: np.ndarray) -> torch.Tensor:
        """Preprocess video for model input"""
        # Implement video preprocessing
        pass
    
    def _preprocess_audio(self, audio: np.ndarray) -> torch.Tensor:
        """Preprocess audio for model input"""
        # Implement audio preprocessing
        pass
    
    def _postprocess_predictions(
        self,
        predictions: Dict[str, np.ndarray]
    ) -> Dict[str, np.ndarray]:
        """Post-process model predictions"""
        # Implement prediction post-processing
        pass 