requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
tensorrt>=8.0.0
opencv-python>=4.5.3
spdlog>=1.9.0

# Distributed processing
aiokafka>=0.7.2
aioredis>=2.0.0
prometheus-client>=0.11.0

# AI and ML
scikit-learn>=0.24.2
scipy>=1.7.1
pandas>=1.3.0

# Security
cryptography>=3.4.7
pyjwt>=2.1.0
python-jose>=3.3.0

# Testing and development
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
mypy>=0.910

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2

# Monitoring and logging
prometheus-client>=0.11.0
python-json-logger>=2.0.2

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.1
tqdm>=4.62.2 